pyproject.toml
[tool.poetry]
name = "drf-openbudget"
version = "0.1.0"
description = ""
authors = ["xolmomin <xolmomin@gmail.com>"]
readme = "README.md"
packages = [{ include = "root" }]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
django-ckeditor = "^6.5.1"
django-mptt = "^0.14.0"
pillow = "^9.5.0"
drf-yasg = "^1.21.5"
django-storages = "^1.13.2"
boto3 = "^1.26.137"
djangorestframework-simplejwt = "^5.2.2"
django-parler = "^2.3"
django-parler-rest = "^2.2"
isort = "^5.12.0"
pytest-django = "^4.5.2"
google-api-python-client = "^2.87.0"
django-redis = "^5.2.0"
dj-database-url = "^2.0.0"
psycopg2-binary = "^2.9.6"
gunicorn = "^20.1.0"
celery = "^5.2.7"
flower = "^1.2.0"
django-filter = "^23.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
skip = "migrations"

# pyproject.toml
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "root.settings"
python_files = "tests.py test_*.py *_tests.py"
